+++ src/Agenda.AppHost/AppHost.cs
@@ -1,8 +1,5 @@
-using Aspire.Hosting;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
 using Projects;
-
 
 IDistributedApplicationBuilder builder = DistributedApplication.CreateBuilder(args);
 
@@ -20,6 +17,9 @@ if (! isRunningIntegrationTests)
             .WithPgWeb(containerName: "pg-web");
 }
 
+var messaging = builder.AddRabbitMQ("messaging")
+    .WithManagementPlugin();
+
 var migrationService = builder.AddProject<Agenda_Migrator>("migrations")
     .WithReference(postgres)
     .WaitFor(postgres);
@@ -29,9 +29,10 @@ var api = builder.AddProject<Agenda_API>("api")
     .WithHttpEndpoint(name: "unsecured")
     .WithHttpsEndpoint(name: "secured")
     .WithReference(postgres)
+    .WithReference(messaging)
+    .WaitFor(messaging)
     .WaitForCompletion(migrationService)
     .PublishAsDockerFile();
-
 builder.Build().Run();