+++ src/Agenda.DataStores.Postgres/Migrations/20251215101906_AddOutboxTable.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using Agenda.DataStores;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using NodaTime;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
@@ -12,9 +13,11 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace Agenda.DataStores.Postgres.Migrations
 {
     [DbContext(typeof(AgendaDataStore))]
-    partial class AgendaDataStoreModelSnapshot : ModelSnapshot
+    [Migration("20251215101906_AddOutboxTable")]
+    partial class AddOutboxTable
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder