+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### 🚀 New features
 - Added `DELETE /appointments/{id}/attendees/{id}` endpoint
+- Added RabbitMQ integration and related configuration.
+- Added publication of `AppointmentScheduled` event when a new appointment is scheduled : 
 
 ### 🧹 Housekeeping
 - Added task issue template
@@ -15,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Codecov configuration file
 - Updated `xRetry` to `1.0.0-rc2`
 - Updated `xunit.v3` to `3.2.0`
+- Updated `Paramore.Brighter.*` packages to `10.0.4`
+- Bumped TFM to `net10.0`
 
 ## [0.1.0] / 2025-10-03
 ### 🚀 New features